fix and cleanup lathe ui
